# openapi.yaml
openapi: 3.0.0
info:
  title: BRDG JPMP API
  description: API for the BRDG JPMP kiosk application
  version: 1.0.0
servers:
  - url: https://api.jpmp.brdg.jarv.us
    description: Production backend
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /items/kiosk/{type}:
    get:
      summary: Get items by kiosk type
      description: Returns a list of items for a specific kiosk type
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum: [sweet, juice]
      responses:
        "200":
          description: A list of items for the specified kiosk type
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Item"
        "400":
          description: Invalid kiosk type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /kiosks:
    get:
      summary: Get all kiosks
      description: Returns a list of all kiosks
      responses:
        "200":
          description: A list of kiosks
          content:
            application/json:
              schema:
                type: object
                properties:
                  kiosks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Kiosk"
    post:
      summary: Create a new kiosk
      description: Creates a new kiosk. When role is 'fulfill', client_kiosk_id is required. When role is not 'fulfill', client_kiosk_id must not be provided.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
                - role
              properties:
                type:
                  type: string
                  enum: [sweet, juice]
                role:
                  type: string
                  enum: [order, fulfill, customize]
                  description: Role of the kiosk
                enabled:
                  type: boolean
                  default: true
                nickname:
                  type: string
                  description: Optional display name for the kiosk
                client_kiosk_id:
                  type: integer
                  description: Reference to another kiosk. Required when role is 'fulfill', not allowed otherwise.
                  nullable: true
                app_version:
                  type: string
                  description: Version of the kiosk application
                  nullable: true
                app_platform:
                  type: string
                  description: Platform running the kiosk application
                  nullable: true
      responses:
        "201":
          description: Kiosk created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  kiosk:
                    $ref: "#/components/schemas/Kiosk"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /kiosks/{id}:
    get:
      summary: Get a specific kiosk
      description: Returns a specific kiosk by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The requested kiosk
          content:
            application/json:
              schema:
                type: object
                properties:
                  kiosk:
                    $ref: "#/components/schemas/Kiosk"
                  currentOrder:
                    type: object
                    nullable: true
                    description: The oldest pending order for this kiosk, or null if none exists
                    allOf:
                      - $ref: "#/components/schemas/Order"
        "404":
          description: Kiosk not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      summary: Update a kiosk
      description: Updates an existing kiosk. When role is 'fulfill', client_kiosk_id is required. When role is not 'fulfill', client_kiosk_id must not be provided.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [sweet, juice]
                  description: Type of kiosk
                role:
                  type: string
                  enum: [order, fulfill, customize]
                  description: Role of the kiosk
                enabled:
                  type: boolean
                nickname:
                  type: string
                  description: Optional display name for the kiosk
                client_kiosk_id:
                  type: integer
                  description: Reference to another kiosk. Required when role is 'fulfill', not allowed otherwise.
                  nullable: true
                app_version:
                  type: string
                  description: Version of the kiosk application
                  nullable: true
                app_platform:
                  type: string
                  description: Platform running the kiosk application
                  nullable: true
      responses:
        "200":
          description: Kiosk updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  kiosk:
                    $ref: "#/components/schemas/Kiosk"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Kiosk not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete a kiosk
      description: Deletes an existing kiosk
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Kiosk deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "404":
          description: Kiosk not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Item:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
          description: Unique identifier for the item
        kiosk_type:
          type: string
          enum: [sweet, juice]
        item_type:
          type: string
          enum: [sweet, juice, gift]
        name:
          type: string
        description:
          type: string
        available:
          type: boolean
        created_at:
          type: string
          format: date-time
    Kiosk:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [sweet, juice]
        role:
          type: string
          enum: [order, fulfill, customize]
          description: Role of the kiosk
        nickname:
          type: string
          description: Optional display name for the kiosk
        enabled:
          type: boolean
        client_kiosk_id:
          type: integer
          description: Reference to another kiosk. Required when role is 'fulfill', not allowed otherwise.
          nullable: true
        app_version:
          type: string
          description: Version of the kiosk application
          nullable: true
        app_platform:
          type: string
          description: Platform running the kiosk application
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          description: Automatically updated when any field changes
    Order:
      type: object
      properties:
        id:
          type: integer
        kiosk_id:
          type: integer
        kiosk_type:
          type: string
          enum: [sweet, juice]
          description: The type of kiosk at the time the order was created. This value is captured when the order is created and does not change even if the kiosk's type changes later.
        status:
          type: string
          enum: [pending, completed, canceled]
        user_profile:
          type: object
          description: User profile data
        survey_response:
          type: object
          description: Survey response data as JSON object
          nullable: true
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        order_id:
          type: integer
        item_id:
          type: integer
        slug:
          type: string
          description: Unique identifier for the item
        name:
          type: string
        description:
          type: string
        available:
          type: boolean
          description: Whether the item is currently available for ordering
        created_at:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            status:
              type: integer